+++ tooling/test-utils/src/accessibility.tsx
@@ -11,7 +11,9 @@ export async function testA11y(
   options: RenderOptions & { axeOptions?: JestAxeConfigureOptions } = {}
 ) {
   const { axeOptions, ...rest } = options
-  const container = React.isValidElement(ui) ? render(ui, rest).container : ui
+  const container = React.isValidElement(ui)
+    ? render(ui, rest).container
+    : (ui as HTMLElement)
   const results = await axe(container, axeOptions)
   expect(results).toHaveNoViolations()
 }