+++ packages/saas-ui-modals/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @saas-ui/modals
 
+## 2.1.31
+
+### Patch Changes
+
+- Updated dependencies [b110d4a]
+  - @saas-ui/forms@2.6.1
+
+## 2.1.30
+
+### Patch Changes
+
+- Updated dependencies [9fe1899]
+  - @saas-ui/forms@2.6.0
+
 ## 2.1.29
 
 ### Patch Changes